Refactored the block code
