--- src/save.js
@@ -1,3 +1,0 @@
-export default function save() {
-    return null;
-}
