--- webpack.config.js
@@ -1,23 +1,0 @@
-const defaultConfig = require('@wordpress/scripts/config/webpack.config');
-
-module.exports = {
-    ...defaultConfig,
-    module: {
-        ...defaultConfig.module,
-        rules: [
-            {
-                test: /\.js$/,
-                exclude: /node_modules/,
-                use: {
-                    loader: 'babel-loader',
-                    options: {
-                        presets: ['@babel/preset-env', '@babel/preset-react']
-                    }
-                }
-            }
-        ],
-    },
-    plugins: [
-        ...defaultConfig.plugins
-    ],
-};
